+++ configurations/hardhat/src/index.ts
@@ -1,4 +1,3 @@
-/* eslint-disable turbo/no-undeclared-env-vars */
 import fs, { readFileSync } from "fs";
 
 import * as dotenv from "dotenv";
@@ -37,11 +36,11 @@ subtask(TASK_COMPILE_SOLIDITY_GET_SOURCE_PATHS).setAction(
   },
 );
 
-let foundry = toml.parse(readFileSync("../../foundry.toml").toString());
+let foundry = toml.parse(readFileSync("foundry.toml").toString());
 
 const getRemappings = () => {
   return fs
-    .readFileSync("../../remappings.txt", "utf8")
+    .readFileSync("remappings.txt", "utf8")
     .split("\n")
     .filter(Boolean)
     .map(line => {
@@ -107,17 +106,16 @@ const config: HardhatUserConfig = {
     },
   },
   paths: {
-    root: "../..",
     artifacts: "packages/contracts/artifacts",
     cache: "packages/contracts/cache",
     deploy: "packages/contracts/deploy",
     deployments: "packages/contracts/deployments",
     imports: "packates/contracts/imports",
     sources: "contracts",
-    tests: "packages/contracts/tests",
+    tests: "contracts/**/spec",
   },
   abiExporter: {
-    path: "./abi",
+    path: "packages/contracts/abi",
     clear: true,
   },
   gasReporter: {
@@ -158,7 +156,7 @@ const config: HardhatUserConfig = {
   watcher: {
     compile: {
       tasks: ["compile"],
-      files: ["./src"],
+      files: ["contracts"],
       verbose: true,
     },
   },
